merged, fungsi 90%

+++ controllers/ebooksController.js
@@ -3,43 +3,51 @@ const { Op } = require('sequelize');
 
 class Controller {
   static findAll(req, res) {
-    let search = req.query.search || '';
-    let email = req.session.email || null;
-    let session;
-    Account.findOne({ where: { email: email } })
-      .then(data => {
-        req.app.locals.session = data;
-        session = req.app.locals.session
-        if (session.role == 'reader') {
-          return EBook.findAll({
-            where: {
-              title: {
-                [Op.iLike]: `%${search}%`
+    if (req.session.isLogin) {
+      let search = req.query.search || '';
+      let email = req.session.email || null;
+      let session;
+      Account.findOne({ where: { email: email } })
+        .then(data => {
+          req.app.locals.session = data;
+          session = req.app.locals.session
+          if (session.role == 'reader') {
+            return EBook.findAll({
+              where: {
+                title: {
+                  [Op.iLike]: `%${search}%`
+                },
+                copies: {
+                  [Op.gt]: 0
+                }
               },
-              copies: {
-                [Op.gt]: 0
-              }
-            },
-            order: ['id']
-          })
-        } else {
-          return EBook.findAll({
-            where: {
-              title: {
-                [Op.iLike]: `%${search}%`
-              }
-            },
-            order: ['id']
-          })
-        }
-      })
-      .then(data => res.render('ebooks', { data, session }))
-      .catch(err => res.send(err));
+              order: ['id']
+            })
+          } else {
+            return EBook.findAll({
+              where: {
+                title: {
+                  [Op.iLike]: `%${search}%`
+                }
+              },
+              order: ['id']
+            })
+          }
+        })
+        .then(data => res.render('ebooks', { data, session }))
+        .catch(err => res.send(err));
+    } else {
+      res.redirect('/');
+    }
   }
 
   static getAdd(req, res) {
-    let session = req.app.locals.session || null;
-    res.render('ebook-form', { data: null, session });
+    if (req.session.isLogin) {
+      let session = req.app.locals.session || null;
+      res.render('ebook-form', { data: null, session });
+    } else {
+      res.redirect('/');
+    }
   }
 
   static postAdd(req, res) {
@@ -49,10 +57,14 @@ class Controller {
   }
 
   static getEdit(req, res) {
-    let session = req.app.locals.session || null;
-    EBook.findByPk(req.params.id)
-      .then(data => res.render('ebook-form', { data, session }))
-      .catch(err => res.send(err));
+    if (req.session.isLogin) {
+      let session = req.app.locals.session || null;
+      EBook.findByPk(req.params.id)
+        .then(data => res.render('ebook-form', { data, session }))
+        .catch(err => res.send(err));
+    } else {
+      res.redirect('/');
+    }
   }
 
   static postEdit(req, res) {
@@ -62,39 +74,51 @@ class Controller {
   }
 
   static details(req, res) {
-    let session = req.app.locals.session || null;
-    EBook.findByPk(req.params.id, {
-        include: [Account],
-        through: {
-          attributes: ['createdAt', 'return_date']
-        }
-      })
-      .then(data => res.render('ebook-details', { data, session }))
-      .catch(err => res.send(err));
+    if (req.session.isLogin) {
+      let session = req.app.locals.session || null;
+      EBook.findByPk(req.params.id, {
+          include: [Account],
+          through: {
+            attributes: ['createdAt', 'return_date']
+          }
+        })
+        .then(data => res.render('ebook-details', { data, session }))
+        .catch(err => res.send(err));
+    } else {
+      res.redirect('/');
+    }
   }
 
   static destroy(req, res) {
-    EBook.destroy({ where: req.params })
-      .then(() => res.redirect('/ebooks'))
-      .catch(err => res.send(err));
+    if (req.session.isLogin) {
+      EBook.destroy({ where: req.params })
+        .then(() => res.redirect('/ebooks'))
+        .catch(err => res.send(err));
+    } else {
+      res.redirect('/');
+    }
   }
 
   static borrow(req, res) {
-    // let email = req.session.email || null;
-    // let account;
-    let session = req.app.locals.session || null;
-    EBook.decrement('copies', { where: req.params })
-      // .then(() => Account.findOne({ where: { email: email } }))
-      .then(() => {
-        // account = data;
-        return borrow_log.create({
-          AccountId: session.id,
-          EBookId: req.params.id,
-          return_date: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000)
-        });
-      })
-      .then(() => res.redirect('/ebooks'))
-      .catch(err => res.send(err));
+    if (req.session.isLogin) {
+      // let email = req.session.email || null;
+      // let account;
+      let session = req.app.locals.session || null;
+      EBook.decrement('copies', { where: req.params })
+        // .then(() => Account.findOne({ where: { email: email } }))
+        .then(() => {
+          // account = data;
+          return borrow_log.create({
+            AccountId: session.id,
+            EBookId: req.params.id,
+            return_date: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000)
+          });
+        })
+        .then(() => res.redirect('/ebooks'))
+        .catch(err => res.send(err));
+    } else {
+      res.redirect('/');
+    }
   }
 }
